CommonFile上传逻辑

+++ xiaohe66-web-portal/src/main/java/com/xiaohe66/web/code/file/po/CommonFile.java
@@ -22,7 +22,7 @@ public class CommonFile extends BasePo {
     // todo : rename to filePath
     protected String fileUrl;
 
-    protected Integer fileByte;
+    protected Long fileByte;
 
     protected Date startTime;
